refactor(protocol): extract opportunity card presentation
