+++ packages/protocol/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@indexnetwork/protocol",
-  "version": "6.13.5",
+  "version": "6.13.6",
   "type": "module",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
